Add docstrings to explain custom methods &  refactor helper
